[dotnet-dev-certs] Report logs from CertificateManager on stdout/stderr
In commit 8e1e81ae78e876eecad1b816093f0519682e205d it was mentioned:
>  We can decide to write an EventListener in `dotnet-dev-certs` to write the results to the console output.

This is what is achieved with the new `ReporterEventListener` in order to bring back ASP.NET Core v3 feature parity where the --verbose option prints debug messages to help diagnose potential certificate issues. 

Also, all the methods from the CertificateManagerEventSource have been appropriately adapted as per the [EventAttribute][1] documentation. Note that it is not required to test for `IsEnabled()` before calling `WriteEvent()` because the very first thing that all `WriteEvent()` overloads do is to return early if not enabled.

[1]: https://docs.microsoft.com/en-us/dotnet/api/system.diagnostics.tracing.eventattribute