[Blazor] Fix publish with different combinations of SWA base paths (#28539)
Description
#28017

In 5.0.1 we fixed a publish scenario where Blazor webassembly wasn't respecting the StaticWebAssetBasePath defined for the project. Unfortunately the logic for handling the path composition was not robust enough and while it fixed the publish output layout, it introduced an integration issue later on at runtime.

The current fix makes the handling of the path more tolerant to initial and final slashes and includes a new end to end test that validates this scenario.

Customer Impact
Customers can't host blazor applications outside the root path (/)

Regression?
Yes. Worked in 3.2

Risk
Low.

We've included additional tests and E2E automation to verify this case.

Validation
 Automated
 Manual