Add an option to tag helper providers to allow filtering (#29688)
As part of converting to source generators, one of the features we miss from MSBuild is the ability to no-op and cache
results operations if inputs are unchanged (target incrementalism). In this particular case, we're trying to
caching tag helper descriptors instead of discovering it every build.

One fairly simple solution is to cache descriptors for reference assemblies (that typically remain unchanged between compilations),
and the current assembly separately. The former can be cached and invalidated when metadatareferences change while the latter has
a smaller set of types to search.