Added guidance about community contribution PRs (#30180)
Updated `Contributing code and content` section to include information about how to choose issues to contribute to and what process to follow.

This is only a part of the workflow. The next one is to make sure we ask community contributors to follow this guidance when they submit a new issue, which I will handle soon.