[Blazor] Adds support for preserving prerendered state in Blazor applications (#29717)
Adds support for preserving and restoring state on Blazor Server and Blazor WebAssembly applications via a new set of primitives.

Components and services in a Blazor application can inject ComponentApplicationState to persist their private state after the application has prerendered and its about to be stopped.

ComponentApplicationState offers methods to save binary data as well as utility methods to persist data as JSON.

Components and services can register an event handler for OnPersisting that is invoked when the app is shutting down and that enables them to save state.

When the application restarts after prerendering, the state is made available to all components during the initial render. After that initial render the state is no longer available.

Redeeming the state for a component is a one time operation, once a component recovers its saved state, the state can no longer be retrieved by any other component or service.

The key used to store the state of a given component or service must be unique across all the components and services that persist data during prerendering.

As a result of the changes required for this feature, the HTML renderer no longer disposes the components inmediately after the initial render, and instead it waits until the end of the request scope to dispose all components.