Suppress generating source checksums when configured
Hot reload / EnC does not like it when type level attributes are modified. Razor views and Pages
include a RazorSourceChecksumAttribute that includes a checksum of all of it's inputs (current cshtml file, and
all _ViewImports that contribute to it). It's used used by runtime compilation to tell if the compiled view is
current compared to it's inputs. However, it gets in the way with enc since editing a file updates the checksum.

We'll disable this feature by default in RazorSourceGenerator, and enable it using an MSBuild switch that's configured
by runtime compilation