Infer default sources for parameters of minimal actions (#31603)
This does a big chunk of #30248 while heavily refactoring the RequestDelegateFactory by splitting it out into a bunch of smaller methods.

For part 2 of #30248, a "scalar value type that lives in the System namespace", I went with the alternative of "Something convention-based like the presence of a bool TryParse(string, out T) method" instead. This conventions work for types in any namespace including user-defined types.