Add support for default parameter values in ViewComponent tag helper (#32979)
Addresses https://github.com/dotnet/aspnetcore/issues/5011.

To implement support for optional parameters, we largely needed to do two things:

- Update the `ViewComponentTagHelperDescriptorFactory` that is responsible for constructing `VCTHD` objects to _not_ produce the `RequiredAttributeDescriptors` for parameters that have a default value provided.
- Update the generated code for VCTHs to produce an arguments object to the `InvokeCoreAsync` method that contains only the provided parameters

**Updates to the VCTH descriptor factory**

The VCTH descriptor contains a `RequiredAttributeDescriptor` for every parameter to the `InvokeAsync` method on the VC. When a parameter has a default value provided, we don't need to configure the parameter as a `RequiredAttribute` on the VCTH descriptor.

With this change in place, instances of the VCTH that don't contain the required attribute will still match with the VCTH. For example, the following VC:


```csharp
public class FooViewComponent : ViewComponent
{
    public IViewResult InvokeAsync(string foo, string bar = "default value")
    {
        return $"{foo} : {bar}";
    }
}
```

Would produce a VCTH descriptor that looks like this:

```
- ViewComponentTagHelper
  - RequiredAttributeDescriptor(propertyName = "foo")
  - BoundAttributeDescriptor(propertyName = "foo")
  - BoundAttributeDescriptor(propertyName = "bar")
```

And will match VCTH instances like:

```
<vc:foo foo="a-foo-value"></vc:foo>
<vc:foo foo="a-foo-value" bar="a-bar-value"></vc:foo>
```

But not VCTH instances that looks like:

```
<vc:foo bar="a-bar-value"></vc:foo>
<vc:foo></vc:foo>
```

All this can happen with very few changes to the tag helper matching logic that already exists.

**Update the generated code for VCTHs**

Presently, whenever a VCTH is encountered, generated code is produced that:

- Writes a class representing the view component which contains properties for each parameter provided to the `InvokeAsync` method and a `ProcessAsync` method that passes the class properties as arguments to `InvokeAsync`
- Instantiates an instance of the class above and sets class properties with their associated attributes

Even if the VCTH is matched to the VC successfully as a result of the changes above, the generated code will be inaccurate because null values will be passed for parameters that have default values. Instead, we want to construct an arguments object that only contains the parameters that were provided as attributes to the VCTH instance.

We do this by adding a new `ProcessInvokeAsyncArgs` that evaluates if each of the parameters was provided as an attribute to the VCTH instance and if so, sets its value in the arguments object.

```csharp
private Dictionary<string, object> ProcessInvokeAsyncArgs(Microsoft.AspNetCore.Razor.TagHelpers.TagHelperContext __context)
{
    Dictionary<string, object> args = new();
    if (__context.AllAttributes.ContainsName("first-name"))
    {
        args[nameof(firstName)] = firstName;
    }
    return args;
}
```

The net effect is that we only pass set parameters as arguments to the method. For parameters that aren't passed as arguments, the default value will be used.