QueryStringEnumerable API (#33910)
* Add QueryStringEnumerable

* Use QueryStringEnumerable in QueryFeature

* Move decoding logic into QueryStringEnumerable

* Simplify parsing logic. Add test and support for another edge case.

Previously, the logic would trim leading whitespace from keys, but only for keys with a value. I don't know why we did this, why we didn't trim trailing whitespace, and why we didn't do it for keys with empty value. Also it was inconsistent - we did it in QueryHelper, but not in QueryFeature.

* Make into a public API

* More unit test cases

* Typo

* Add unit test to clarify QueryHelpers whitespace handling

* Fix access modifier

* CR: Remove unwanted test

* Optimize: avoid allocations in common cases