[Static web assets] Updates hosting to use the new static web assets manifest format (#34510)
Updates static web assets to use the new JSON based manifest format.

The build now produces a <<AppName>>.staticwebasets.runtime.json that
contains a list of content roots as well as a tree structure that
includes the paths for the list of mapped static web assets as well as
the list of patterns to use for discovering new asses.