Support reloadable cshtml (#35017)
As part of hot reload, types marked with `CreateNewOnMetadataUpdateAttribute` will result in a new type created rather than updating existing types. This works really well for .cshtml scenario where the type is never referenced by name, but rather using the path. This change adds the annotation to the cshtml file and does the runtime work to substitute the build-time compiled file with hot reloaded type on a hot reloaded event. 

* Annotate cshtml files with `CreateNewOnMetadataUpdateAttribute` when targeting RazorLangVersion = 6.0
* Additionally annotate cshtml files with an `RazorCompiledItemMetadataAttribute attribute that identifies the path (identifier)
* Update RazorCompiledItemFeatureProvider to substitute a compiled type with a hot reloaded type
* Clear more Razor specific caches as part of a hot reload event

Fixes #34248