Avoid using new C# language features in Razor code generation
The generated code for tag helpers is used all the way back in 3.1 which defaults to C# 8. Our code generation
currently produces a target-typed new expression that is available in C# 10.

Updating some of the older tests to pin the language to a fixed version so we can catch these kinds of regressions as part of tests.

Fixes https://github.com/dotnet/aspnetcore/issues/35554