Always download blazor-hotreload.js from app root
The middleware that we inject always listens to the root of the app. While this might not play very nicely if the app
is running off a virtual path, listening to the root is what we do with the aspnet-browser-refresh.js script and we've had no issue
reports with it thus far.

Fixes https://github.com/dotnet/aspnetcore/issues/35555