Fix DatabaseDeveloperPageExceptionFilter to handle wrapped DbExceptions
Sometimes EF (or other libraries) wrap database errors. The exception filter should account for this, but was not doing. This was revealed by https://github.com/dotnet/efcore/issues/25050 where we started treating more error numbers as transient and hence wrapping their exceptions.

Note that the original Diagnostics.EFCore.FunctionalTests have a test for this, but it appears that these tests were never updated when the mechanism was changed in .NET 5.