Revert consolidating DefaultViewCompilerProvider + DefaultViewCompiler (#37037)
As part of https://github.com/dotnet/aspnetcore/pull/35017, I had consolidated the DefaultViewCompiler and DefaultViewCompilerProvider
types, largely as a clean up. The type DefaultViewCompilerProvider is looked up by the RuntimeCompilation package by name, and when making
this change, this code was also updated.

However, we've had two reports so far of users trying to use an 5.0 or earlier version of the runtime compilation package with a 6.0 app. In this case, the lookup
fails and the runtime compilation feature doesn't work as expected after this.

Fixes https://github.com/dotnet/aspnetcore/issues/35267