Avoid inferring RequiredAttribute for properties of types in the framework
* Validation in MVC is recursive, but will short-circuit if none of the sub-properties on a given instance require validation.

* In 3.1, MVC added a feature where it will annotate non-nullable properties and parameters with a Required attribute in it's model metadata. Validation
would then validate that the property was assigned a non-null value.

In tandem, the 2nd bullet point makes validation perform very poorly for framework or library types that have been annotated with non-null attributes. In particular,
when the nullability does not describe what the outcome should be when binding a model, but rather the nullability behavior of a well-formed instance. A very tactical
change we could make here is to avoid annotating properties on types declared in the framework with the implicit [Required].

It's important to know that the implicit Required feature does not apply to generic types (MVC does not know how to interpret nullability + generic types), so we don't have
a scenario where a framework type could hold a reference to a user type at compile time (e.g List<Person>).

Fixes https://github.com/dotnet/aspnetcore/issues/36919