[Static web assets] Fixes an issue where patterns at the root were not being considered (#37560) (#37580)
## Description

Static web assets falls back to check on a set of defined folders when it can't find a file in its manifest at runtime. This allows us to map new and renamed files while the app without having to restart the app, in the same way it works when you are dealing with a web project and content in the wwwroot folder.

Blazor Webassembly leverages this capability to allow the user to add and rename files to the client wwwroot folder and automatically serve them on a page refresh without having to restart the application.

Static web assets contains a list of all the potential locations for content in its manifest (simplified) as a triplet of (base path, pattern, folder).

The issue is that Blazor Webassembly applications typically serve these files from the root of the url space "/" and when we fail to find a file on the manifest and need to check the pattern ("/", "**", <<folder>>) added by the Blazor webassembly app, we don't process the path correctly to remove the initial "/' from the path, and that causes the pattern matching to not match.

The fix initializes the depth to "-1" instead of "0" which forces the algorithm to go through the same steps that it follows for other pattern and normalize the path before the evaluation against the pattern. This way, a request like "/other.html" is preprocessed to "other.html" which matches the pattern "**" and causes the file to be returned if present.

Fixes #37508

## Customer Impact

Customers can't add or rename a file in their blazor webassembly app project wwwroot and have it served without stopping and rebuilding the app. This puts a dent on the inner loop performance

## Regression?

- [X] Yes
- [ ] No

5.0

## Risk

- [ ] High
- [ ] Medium
- [X] Low

Its a bug in the initialization conditions of the algorithm, we added a specific test to cover this case.

## Verification

- [X] Manual (required)
- [X] Automated

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [X] N/A