Pass IJSVoidResult instead of object to Invoke to indicate void result (#37532) (#37566)
# Pass `IJSVoidResult` instead of `object` to `InvokeVoid` to indicate void result

Change `InvokeVoid` in `JSInProcessObjectReferenceExtensions` & `JSInProcessRuntimeExtensions` to have a generic return value of `IJSVoidResult` instead of `object` to indicate void result.

## Description

As part of https://github.com/dotnet/aspnetcore/issues/35929, we moved from `InvokeVoidAsync<object>` to `InvokeVoidAsync<IJSVoidResult>`. However, this change is missing for `InvokeVoid`. This PR updates the `InvokeVoid` APIs to be in line with `InvokeVoidAsync`, to ensure operational consistency and decrease interop overhead. 

Fixes #37528

## Customer Impact

Without this change, `InvokeVoid` and `InvokeVoidAsync` behave differently, which may lead to user confusion. This also [causes issues](https://github.com/dotnet/aspnetcore/issues/37528) with testability in external testing frameworks like [`bunit`](https://bunit.dev/). 

Additionally, without this change if users attempt to return non-serializable values like a JavaScript `window`, then they will see an error with `InvokeVoid` but not with `InvokeVoidAsync`.

Currently, we're serializing the return value of `InvokeVoid`, and then discarding it. With this change, we do not serialize the return value at all. Consequently, this change has performance benefits as well.

## Regression?

- [ ] Yes
- [x] No

## Risk

- [ ] High
- [ ] Medium
- [x] Low

We discard the return value of `InvokeVoid` regardless, by setting the generic return value to `IJSVoidResult` instead of `object`, we ensure the return value isn't serialized in the first place. 

## Verification

- [x] Manual (required)
- [x] Automated

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [x] N/A