Update README.md
Fixes: https://github.com/dotnet/aspnetcore/issues/37654

Updated E2E test execution instructions based on what we're currently doing in the CI:

https://github.com/dotnet/aspnetcore/blob/372de32302eecce686fa40cf5cc02d293434ddc7/.azure/pipelines/components-e2e-tests.yml#L38-L56

Note, `./src/Components/build.cmd -test` isn't sufficient in this case as it leads to build errors (it's wanting stuff from `src/Identity` and so on). Doing a root level `./build.cmd` first may resolve it, but would want to avoid that whenever possible as that can be an intensive process.