Avoid null ref in TagHelperMatchingConvention when attribute name is null. (#38264) (#38271)
As part of .NET 6, we made some  tweaks to the razor parser to avoid string allocations. This change
results in a null-ref for bound properties without public setters. The parser change affects building
apps going back all the way to .NET 2.1, so we'd like to fix this.

Fixes https://github.com/dotnet/aspnetcore/issues/38194