Allow runtime compilation to be configured via the hosting startup. (#39381)
As part of refactoring how Razor Pages are loaded, the ability to use runtime compilation via the hosting startup / env variable was broken.
The refactoring assumed AddRuntimeCompilation was always invoked after a call to AddRazorPages / AddControllersWithViews etc which is not the case
for registration via the HostingStartup. This PR address this defect.

Fixes https://github.com/dotnet/aspnetcore/issues/38465