Do not upgrade warnings to errors in VS (#39845)
Now that we have enabled a bunch of code analysis, we've discovered it adds friction to the dev process. This change aims to reduce the friction by preventing warnings to be upgraded to errors in VS. This way you can continue developing, while possible issues are still surfaced. E.g. from a recent error: