Don't compare SequencePosition instances
We can'tompare SequencePosition instances. The problem is that consumed could theoretically point to the very end of a ReadOnlySequenceSegment<byte> and _readResult.Buffer.End might point to the very beginning of the next ReadOnlySequenceSegment<byte>. All the bytes would be consumed in this case, but the SequencePositions wouldn't be equal.