Added the FabricBot configuration file (#40278)
Fabric Bot now supports code-based configuration. That means, as soon as this change is merged, Fabric Bot will be using this configuration going forward, instead of the one in the portal.

For now the values in this file can't reference any variables known to the build system - Version.props.
But I've reached out to the FabricBot team and they've agreed that the scenario makes sense and added [an item to their backlog](https://mseng.visualstudio.com/1ES/_workitems/edit/1919391) to support this in the future.

Also, for now there is no support for per-branch configuration, but support for it is also coming. They're already working on that.

This change also impacts the branching instructions. PR for the change is here: https://github.com/dotnet/aspnetcore-internal/pull/4010