Update to limit phone number length
Update ApplicationDbContextModelSnapshot.cs to limit phone number length. 

Using nvarchar(max) in a nonclustered index is only possible by adding it to some other index key as an include. https://stackoverflow.com/questions/12336821/how-can-i-create-index-on-nvarcharmax-datatype-in-sql

By reducing the length to something a little more reasonable it is possible to create a nonclustered index on phone number directly for better query performance at scale.

The longest phone number is currently 15 digits. I've reduced this field length from NVARCHAR(MAX) to NVARCHAR(32) to allow for adding of indexes if desired and future expansion of digits.

https://en.wikipedia.org/wiki/Telephone_numbering_plan#:~:text=The%20International%20Telecommunication%20Union%20(ITU,15%20digits%20to%20telephone%20numbers.