Modify HtmlRenderer to update content when a component re-renders (#40512)
Prior to this change HtmlRenderer would return a snapshot of a component immediately after it was initially rendered. This meant any updates to a component as a result of other root components rendering were lost. 

In this change, we update the HtmlRenderer to return a placeholder `IHtmlContent` that retrieves render frames only when `WriteTo` is invoked. Typically MVC will defer calling WriteTo until the content is about to be sent to the wire, which means any updates to a component (e.g. HeadOutlet being updated) are picked up.

Also updates the template to consolidate `_Host` and `_Layout`.

Fixes https://github.com/dotnet/aspnetcore/issues/38400